Upgrade to csdk 2.0 - Bump CSDK, Next.js, and Node type versions according to release note
