+++ examples/kit-nextjs-product-listing/.sitecore/component-map.client.ts
@@ -16,6 +16,7 @@ import * as TestimonialCarousel from 'src/components/testimonial-carousel/Testim
 import * as NavigationMenuToggleclient from 'src/components/sxa/NavigationMenuToggle.client';
 import * as NavigationListclient from 'src/components/sxa/NavigationList.client';
 import * as LinkList from 'src/components/sxa/LinkList';
+import * as ContentBlock from 'src/components/sxa/ContentBlock';
 import * as ButtonNavigationclient from 'src/components/sxa/ButtonNavigation.client';
 import * as SubscriptionBanner from 'src/components/subscription-banner/SubscriptionBanner';
 import * as SlideCarouseldev from 'src/components/slide-carousel/SlideCarousel.dev';
@@ -137,6 +138,7 @@ export const componentMap = new Map<string, NextjsContentSdkComponent>([
   ['NavigationMenuToggle', { ...NavigationMenuToggleclient }],
   ['NavigationList', { ...NavigationListclient }],
   ['LinkList', { ...LinkList }],
+  ['ContentBlock', { ...ContentBlock }],
   ['ButtonNavigation', { ...ButtonNavigationclient }],
   ['SubscriptionBanner', { ...SubscriptionBanner }],
   ['SlideCarousel', { ...SlideCarouseldev }],
