+++ examples/kit-nextjs-product-listing/package.json
@@ -3,12 +3,6 @@
   "description": "Application utilizing Sitecore JavaScript Services and Next.js",
   "version": "1.0.0",
   "private": true,
-  "config": {
-    "appName": "kit-nextjs-product-starter",
-    "graphQLEndpointPath": "/sitecore/api/graph/edge",
-    "language": "en",
-    "template": "nextjs"
-  },
   "author": {
     "name": "Sitecore Corporation",
     "url": "https://doc.sitecore.com/xmc/en/developers/content-sdk/index.html"
@@ -61,7 +55,7 @@
     "@radix-ui/themes": "^3.2.1",
     "@sitecore-cloudsdk/core": "^0.5.4",
     "@sitecore-cloudsdk/events": "^0.5.4",
-    "@sitecore-content-sdk/nextjs": "^1.0.0",
+    "@sitecore-content-sdk/nextjs": "^2.0.0",
     "@sitecore-feaas/clientside": "^0.6.0",
     "@sitecore/components": "~2.1.0",
     "@tailwindcss/cli": "^4.1.11",
@@ -84,7 +78,7 @@
     "input-otp": "^1.4.2",
     "lucide-react": "^0.475.0",
     "motion": "^12.23.0",
-    "next": "^15.5.10",
+    "next": "^16.1.1",
     "next-intl": "^4.3.5",
     "next-themes": "^0.4.6",
     "postcss": "^8.5.6",
@@ -111,13 +105,13 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
-    "@sitecore-content-sdk/cli": "^1.0.0",
+    "@sitecore-content-sdk/cli": "^2.0.0",
     "@tailwindcss/postcss": "^4",
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.3.1",
     "@testing-library/user-event": "^14.6.1",
     "@types/jest": "^30.0.0",
-    "@types/node": "^22.16.0",
+    "@types/node": "^24.10.4",
     "@types/react": "^19.1.10",
     "@types/react-dom": "^19.1.7",
     "@typescript-eslint/eslint-plugin": "^8.35.1",
@@ -126,7 +120,7 @@
     "cross-env": "^10.0.0",
     "dotenv-flow": "^4.1.0",
     "eslint": "^9.33.0",
-    "eslint-config-next": "15.5.0",
+    "eslint-config-next": "16.1.1",
     "eslint-config-prettier": "^8.10.0",
     "eslint-plugin-prettier": "^5.4.0",
     "eslint-plugin-react": "^7.37.5",
