+++ examples/kit-nextjs-product-listing/src/components/sxa/ContentBlock.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Text, RichText, Field, withDatasourceCheck } from '@sitecore-content-sdk/nextjs';
 import { ComponentProps } from 'lib/component-props';
 import { JSX } from 'react';
