+++ examples/kit-nextjs-skate-park/package.json
@@ -2,12 +2,6 @@
   "name": "content-sdk-nextjs-skate-park-app-router",
   "version": "1.0.0",
   "private": true,
-  "config": {
-    "appName": "content-sdk-nextjs-skate-park-app-router",
-    "graphQLEndpointPath": "/sitecore/api/graph/edge",
-    "language": "en",
-    "template": "nextjs"
-  },
   "author": {
     "name": "Sitecore Corporation",
     "url": "https://doc.sitecore.com/xmc/en/developers/content-sdk/index.html"
@@ -23,8 +17,8 @@
   "dependencies": {
     "@sitecore-cloudsdk/core": "^0.5.4",
     "@sitecore-cloudsdk/events": "^0.5.4",
-    "@sitecore-content-sdk/nextjs": "^1.0.0",
-    "next": "^15.5.10",
+    "@sitecore-content-sdk/nextjs": "^2.0.0",
+    "next": "^16.1.1",
     "next-intl": "^4.3.5",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
@@ -33,19 +27,19 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
-    "@sitecore-content-sdk/cli": "^1.0.0",
+    "@sitecore-content-sdk/cli": "^2.0.0",
     "@tailwindcss/postcss": "^4",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.3.0",
     "@types/jest": "^30.0.0",
-    "@types/node": "^22.15.14",
+    "@types/node": "^24.10.4",
     "@types/react": "^19.1.10",
     "@types/react-dom": "^19.1.7",
     "axios": "^1.13.6",
     "cross-env": "^10.0.0",
     "eslint": "^9.33.0",
-    "eslint-config-next": "15.5.0",
+    "eslint-config-next": "16.1.1",
     "fast-xml-parser": "^5.4.2",
     "jest": "^30.2.0",
     "jest-environment-jsdom": "^30.2.0",
