+++ examples/kit-nextjs-skate-park/src/app/[site]/[locale]/[[...path]]/page.tsx
@@ -89,7 +89,14 @@ export const generateStaticParams = async () => {
       routing.locales.slice()
     );
   }
-  return [];
+  // Next.js 16 requires at least one param set to avoid EmptyGenerateStaticParamsError
+  return [
+    {
+      site: (sites[0] as SiteInfo)?.name || 'default',
+      locale: routing.defaultLocale || scConfig.defaultLanguage,
+      path: [],
+    },
+  ];
 };
 
 // Metadata fields for the page.
