+++ examples/kit-nextjs-skate-park/src/app/api/sitemap-llm/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { SiteResolver } from '@sitecore-content-sdk/core/site';
-import type { SitemapXmlOptions } from '@sitecore-content-sdk/core/client';
+import { SiteResolver } from '@sitecore-content-sdk/content/site';
+import type { SitemapXmlOptions } from '@sitecore-content-sdk/content/client';
 import type { SiteInfo } from '@sitecore-content-sdk/nextjs';
 import client from 'lib/sitecore-client';
 import sites from '.sitecore/sites.json';
