+++ examples/kit-nextjs-skate-park/src/components/content-block/ContentBlock.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { JSX } from 'react';
 import { Text, RichText, Field, withDatasourceCheck } from '@sitecore-content-sdk/nextjs';
 import { ComponentProps } from 'lib/component-props';
