+++ xmcloud.build.json
@@ -2,7 +2,7 @@
   "renderingHosts": {
     "nextjsstarter": {
       "path": "./examples/kit-nextjs-skate-park",
-      "nodeVersion": "22.22.0",
+      "nodeVersion": "24",
       "jssDeploymentSecret": "110F1C44A496B45478640DD36F80C18C9",
       "enabled": true,
       "type": "sxa",
@@ -38,7 +38,7 @@
     },
     "kit-nextjs-product-starter": {
       "path": "./examples/kit-nextjs-product-listing",
-      "nodeVersion": "22.22.0",
+      "nodeVersion": "24",
       "jssDeploymentSecret": "110F1C44A496B45478640DD36F80C18C9",
       "enabled": true,
       "type": "sxa",
